Move discord models to accounts app. Keep services in home.
