+++ accounts/admin.py
@@ -16,12 +16,13 @@ from django.urls import reverse
 from accounts.models import (
     ButtondownAccount,
     CustomUser,
+    DiscordMember,
+    DiscordRole,
     Link,
     UserAvailability,
     UserProfile,
 )
 from home.models.session import SessionMembership
-from home.models import DiscordMember
 from indymeet.admin import DescriptiveSearchMixin
 
 
@@ -180,6 +181,46 @@ class UserProfileAdmin(ExportCsvMixin, DescriptiveSearchMixin, admin.ModelAdmin)
         if db_field.name == "discord_member":
             kwargs["queryset"] = DiscordMember.objects.all()
         return super().formfield_for_foreignkey(db_field, request, **kwargs)
+
+
+@admin.register(DiscordRole)
+class DiscordRoleAdmin(admin.ModelAdmin):
+    """Read-only view of the roles an announcement can ping.
+
+    The rows are a mirror of the Discord server, rewritten wholesale by the
+    Discord session setup action and the ``sync_discord_roles`` command, so
+    editing them here would only be undone on the next sync.
+    """
+
+    list_display = ("name", "discord_id", "updated_at")
+    search_fields = ("name",)
+    readonly_fields = ("name", "discord_id", "created_at", "updated_at")
+
+    def has_add_permission(self, request) -> bool:
+        return False
+
+
+@admin.register(DiscordMember)
+class DiscordMemberAdmin(admin.ModelAdmin):
+    """Read-only view of guild members mirrored from Discord."""
+
+    list_display = (
+        "nickname",
+        "username",
+        "discord_id",
+    )
+    search_fields = ("username", "nickname", "discord_id")
+    readonly_fields = (
+        "discord_id",
+        "username",
+        "nickname",
+        "role_ids",
+        "created_at",
+        "updated_at",
+    )
+
+    def has_add_permission(self, request) -> bool:
+        return False
 
 
 @admin.register(ButtondownAccount)
