+++ accounts/tests/test_queryset_managers.py
@@ -5,8 +5,12 @@ from django.contrib.auth.models import Permission
 from django.contrib.contenttypes.models import ContentType
 from django.test import TestCase
 
-from accounts.factories import UserAvailabilityFactory, UserFactory
-from accounts.models import CustomUser
+from accounts.factories import (
+    DiscordMemberFactory,
+    UserAvailabilityFactory,
+    UserFactory,
+)
+from accounts.models import CustomUser, DiscordMember
 from home.factories import (
     OrganizerFactory,
     ProjectFactory,
@@ -315,3 +319,31 @@ class ForComparingAvailabilityTestCase(TestCase):
         user_ids = [u.id for u in users]
 
         self.assertEqual(len(user_ids), len(set(user_ids)))
+
+
+class DiscordMemberQuerySetTestCase(TestCase):
+    """Test DiscordMemberQuerySet methods."""
+
+    @classmethod
+    def setUpTestData(cls):
+        cls.unlinked = DiscordMemberFactory.create()
+        cls.linked = DiscordMemberFactory.create()
+        cls.user = UserFactory.create()
+        cls.user.profile.discord_member = cls.linked
+        cls.user.profile.save(update_fields=["discord_member"])
+
+    def test_unassigned_excludes_linked_members(self):
+        self.assertCountEqual(DiscordMember.objects.unassigned(), [self.unlinked])
+
+    def test_unassigned_or_for_user_includes_own_member(self):
+        self.assertCountEqual(
+            DiscordMember.objects.unassigned_or_for_user(self.user),
+            [self.unlinked, self.linked],
+        )
+
+    def test_unassigned_or_for_user_excludes_other_users_member(self):
+        other_user = UserFactory.create()
+
+        self.assertCountEqual(
+            DiscordMember.objects.unassigned_or_for_user(other_user), [self.unlinked]
+        )
