+++ home/integrations/discord/service.py
@@ -7,8 +7,8 @@ from django.conf import settings
 from django.db.models.signals import post_delete, post_save
 from django.dispatch import receiver
 
+from accounts.models import DiscordRole
 from home.integrations.discord.client import DiscordClient
-from home.models import DiscordRole
 
 logger = logging.getLogger(__name__)
